+++ packages/create-app/template/content-collections/content-collections.ts
@@ -17,6 +17,7 @@ const metas = defineCollection({
   name: 'meta',
   directory: 'content/docs',
   include: '**/meta.json',
+  parser: 'json',
   schema: createMetaSchema,
 });