+++ packages/ui/src/layout.client.tsx
@@ -1,17 +1,23 @@
 'use client';
 
 import { MoreVertical } from 'lucide-react';
-import type { LinkItemType, SharedNavProps } from '@/layout';
+import type { SharedNavProps } from '@/layout';
 import { useSearchContext } from '@/contexts/search';
 import { useI18n } from '@/contexts/i18n';
-import { LinkItem } from '@/components/layout/link-item';
+import {
+  LinkItem,
+  type LinkItemType,
+  LinksMenu,
+} from '@/components/layout/link-item';
 import {
   LargeSearchToggle,
   SearchToggle,
 } from '@/components/layout/search-toggle';
 import { ThemeToggle } from '@/components/layout/theme-toggle';
-import { LinksMenu } from '@/components/layout/links-menu';
 import { NavBox, Title } from '@/components/layout/nav';
+import { cn } from '@/utils/cn';
+import { buttonVariants } from '@/theme/variants';
+import { isSecondary } from '@/utils/shared';
 
 export function Nav({
   title,
@@ -30,13 +36,9 @@ export function Nav({
         <Title title={title} url={url} />
         {children}
         {items
-          .filter((item) => item.type !== 'secondary')
+          .filter((item) => !isSecondary(item))
           .map((item, i) => (
-            <LinkItem
-              key={i}
-              item={item}
-              className="-mx-2 text-sm max-lg:hidden"
-            />
+            <LinkItem key={i} item={item} className="text-sm max-lg:hidden" />
           ))}
         <div className="flex flex-1 flex-row items-center justify-end md:gap-2">
           {enableSearch && search.enabled ? (
@@ -47,15 +49,19 @@ export function Nav({
           ) : null}
 
           <ThemeToggle className="max-lg:hidden" />
-          {items
-            .filter((item) => item.type === 'secondary')
-            .map((item, i) => (
-              <LinkItem key={i} item={item} className="max-lg:hidden" />
-            ))}
+          {items.filter(isSecondary).map((item, i) => (
+            <LinkItem key={i} item={item} className="max-lg:hidden" />
+          ))}
 
           <LinksMenu
             items={items}
-            className="-me-2 lg:hidden"
+            className={cn(
+              buttonVariants({
+                size: 'icon',
+                color: 'ghost',
+                className: '-me-2 lg:hidden',
+              }),
+            )}
             footer={
               <div className="flex flex-row items-center justify-between px-2 pt-2">
                 <p className="font-medium text-muted-foreground">